+++ extra_data/__init__.py
@@ -33,9 +33,10 @@ You should have received a copy of the 3-Clause BSD License along with this
 program. If not, see <https://opensource.org/licenses/BSD-3-Clause>
 """
 
-__version__ = "1.3.0"
+__version__ = "1.4.1"
 
 
+from .keydata import KeyData
 from .reader import *
 from .stacking import *
 from .utils import *