+++ extra_data/components.py
@@ -15,33 +15,6 @@ log = logging.getLogger(__name__)
 MAX_PULSES = 2700
 
 
-def _guess_axes(data, train_pulse_ids, unstack_pulses):
-    # Raw files have a spurious extra dimension
-    if data.ndim >= 2 and data.shape[1] == 1:
-        data = data[:, 0]
-
-    # TODO: this assumes we can tell what the axes are just from the
-    # number of dimensions. Works for the data we've seen, but we
-    # should look for a more reliable way.
-    if data.ndim == 4:
-        # image.data in raw data
-        dims = ['train_pulse', 'data_gain', 'slow_scan', 'fast_scan']
-    elif data.ndim == 3:
-        # image.data, image.gain, image.mask in calibrated data
-        dims = ['train_pulse', 'slow_scan', 'fast_scan']
-    else:
-        # Everything else seems to be 1D
-        dims = ['train_pulse']
-
-    arr = xarray.DataArray(data, {'train_pulse': train_pulse_ids}, dims=dims)
-
-    if unstack_pulses:
-        # Separate train & pulse dimensions, and arrange dimensions
-        # so that the data is contiguous in memory.
-        dim_order = ['train', 'pulse'] + dims[1:]
-        return arr.unstack('train_pulse').transpose(*dim_order)
-    else:
-        return arr
 
 
 def _check_pulse_selection(pulses):
@@ -81,13 +54,13 @@ def _check_pulse_selection(pulses):
     return type(pulses)(val)
 
 
-class MPxDetectorBase:
-    """Base class for megapixel detectors (AGIPD, LPD)
+class MultimodDetectorBase:
+    """Base class for detectors made of several modules as separate data sources
     """
 
     _source_re = re.compile(r'(?P<detname>.+)/DET/(\d+)CH')
-    n_modules = 16
     # Override in subclass
+    _main_data_key = ''  # Key to use for checking data counts match
     module_shape = (0, 0)
 
     def __init__(self, data: DataCollection, detector_name=None, modules=None,
@@ -107,7 +80,7 @@ class MPxDetectorBase:
         # pandas' missing-data handling converts the data to floats if there
         # are any gaps - so fill them with 0s and convert back to uint64.
         mod_data_counts = pd.DataFrame({
-            src: data.get_data_counts(src, 'image.data')
+            src: data.get_data_counts(src, self._main_data_key)
             for src in source_to_modno
         }).fillna(0).astype(np.uint64)
 
@@ -139,7 +112,7 @@ class MPxDetectorBase:
         for source in data.instrument_sources:
             m = cls._source_re.match(source)
             if m:
-                detector_names.add(m.group(1))
+                detector_names.add(m.group('detname'))
         if not detector_names:
             raise SourceNameError(cls._source_re.pattern)
         elif len(detector_names) > 1:
@@ -151,21 +124,19 @@ class MPxDetectorBase:
             )
         return detector_names.pop()
 
-    @staticmethod
-    def _identify_sources(data, detector_name=None, modules=None):
-        detector_re = re.compile(re.escape(detector_name) + r'/DET/(\d+)CH')
+    def _identify_sources(self, data, detector_name, modules=None):
         source_to_modno = {}
         for source in data.instrument_sources:
-            m = detector_re.match(source)
-            if m:
-                source_to_modno[source] = int(m.group(1))
+            m = self._source_re.match(source)
+            if m and m.group('detname') == detector_name:
+                source_to_modno[source] = int(m.group('modno'))
 
         if modules is not None:
             source_to_modno = {s: n for (s, n) in source_to_modno.items()
                                if n in modules}
 
         if not source_to_modno:
-            raise SourceNameError(detector_re.pattern)
+            raise SourceNameError(f'{detector_name}/DET/...')
 
         return source_to_modno
 
@@ -197,6 +168,99 @@ class MPxDetectorBase:
         return "<{}: Data interface for detector {!r} with {} modules>".format(
             type(self).__name__, self.detector_name, len(self.source_to_modno),
         )
+
+    @staticmethod
+    def _concat(arrays, index, fill_value, astype):
+        dtype = arrays[0].dtype if astype is None else np.dtype(astype)
+        if fill_value is None:
+            fill_value = np.nan if dtype.kind == 'f' else 0
+        fill_value = dtype.type(fill_value)
+
+        return xarray.concat(
+            [a.astype(dtype, copy=False) for a in arrays],
+            pd.Index(index, name='module'),
+            fill_value=fill_value
+        )
+
+    def get_array(self, key, *, fill_value=None, roi=(), astype=None):
+        """Get a labelled array of detector data
+
+        Parameters
+        ----------
+
+        key: str
+          The data to get, e.g. 'image.data' for pixel values.
+        fill_value: int or float, optional
+            Value to use for missing values. If None (default) the fill value
+            is 0 for integers and np.nan for floats.
+        roi: tuple
+          Specify e.g. ``np.s_[10:60, 100:200]`` to select pixels within each
+          module when reading data. The selection is applied to each individual
+          module, so it may only be useful when working with a single module.
+        astype: Type
+          Data type of the output array. If None (default) the dtype matches the
+          input array dtype
+        """
+        arrays = []
+        modnos = []
+        for modno, source in sorted(self.modno_to_source.items()):
+            arrays.append(self.data.get_array(source, key, roi=roi))
+            modnos.append(modno)
+
+        return self._concat(arrays, modnos, fill_value, astype)
+
+    def get_dask_array(self, key, fill_value=None, astype=None):
+        """Get a labelled Dask array of detector data
+        Parameters
+        ----------
+        key: str
+          The data to get, e.g. 'image.data' for pixel values.
+        fill_value: int or float, optional
+          Value to use for missing values. If None (default) the fill value is 0
+          for integers and np.nan for floats.
+        astype: Type
+          Data type of the output array. If None (default) the dtype matches the
+          input array dtype
+        """
+        arrays = []
+        modnos = []
+        for modno, source in sorted(self.modno_to_source.items()):
+            modnos.append(modno)
+            mod_arr = self.data.get_dask_array(source, key, labelled=True)
+            arrays.append(mod_arr)
+
+        return self._concat(arrays, modnos, fill_value, astype)
+
+    def trains(self, require_all=True):
+        """Iterate over trains for detector data.
+
+        Parameters
+        ----------
+
+        require_all: bool
+          If True (default), skip trains where any of the selected detector
+          modules are missing data.
+
+        Yields
+        ------
+
+        train_data: dict
+          A dictionary mapping key names (e.g. ``image.data``) to labelled
+          arrays.
+        """
+        return MPxDetectorTrainIterator(self, require_all=require_all)
+
+
+class XtdfDetectorBase(MultimodDetectorBase):
+    """Common machinery for a group of detectors with similar data format
+
+    AGIPD, DSSC & LPD all store pulse-resolved data in an "image" group,
+    with both trains and pulses along the first dimension. This allows a
+    different number of frames to be stored for each train, which makes
+    access more complicated.
+    """
+    n_modules = 16
+    _main_data_key = 'image.data'
 
     @staticmethod
     def _select_pulse_ids(pulses, data_pulse_ids):
@@ -239,7 +303,53 @@ class MPxDetectorBase:
 
         return np.concatenate(positions)
 
-    def _get_module_pulse_data(self, source, key, pulses, unstack_pulses):
+    def _make_image_index(self, tids, inner_ids, inner_name='pulse'):
+        # Overridden in LPD1M for parallel gain mode
+        return pd.MultiIndex.from_arrays(
+            [tids, inner_ids], names=['train', inner_name]
+        )
+
+    @staticmethod
+    def _guess_axes(data, train_pulse_ids, unstack_pulses):
+        # Raw files have a spurious extra dimension
+        if data.ndim >= 2 and data.shape[1] == 1:
+            data = data[:, 0]
+
+        # TODO: this assumes we can tell what the axes are just from the
+        # number of dimensions. Works for the data we've seen, but we
+        # should look for a more reliable way.
+        if data.ndim == 4:
+            # image.data in raw data
+            dims = ['train_pulse', 'data_gain', 'slow_scan', 'fast_scan']
+        elif data.ndim == 3:
+            # image.data, image.gain, image.mask in calibrated data
+            dims = ['train_pulse', 'slow_scan', 'fast_scan']
+        else:
+            # Everything else seems to be 1D
+            dims = ['train_pulse']
+
+        arr = xarray.DataArray(data, {'train_pulse': train_pulse_ids},
+                               dims=dims)
+
+        if unstack_pulses:
+            # Separate train & pulse dimensions, and arrange dimensions
+            # so that the data is contiguous in memory.
+            dim_order = train_pulse_ids.names + dims[1:]
+            return arr.unstack('train_pulse').transpose(*dim_order)
+        else:
+            return arr
+
+    def _get_module_pulse_data(self, source, key, pulses, unstack_pulses,
+                               inner_index='pulseId', roi=()):
+        def get_inner_ids(f, data_slice, ix_name='pulseId'):
+            ids = f.file[f'/INSTRUMENT/{source}/{group}/{ix_name}'][
+                data_slice
+            ]
+            # Raw files have a spurious extra dimension
+            if ids.ndim >= 2 and ids.shape[1] == 1:
+                ids = ids[:, 0]
+            return ids
+
         seq_arrays = []
         data_path = "/INSTRUMENT/{}/{}".format(source, key.replace('.', '/'))
         for f in self.data._source_index[source]:
@@ -259,29 +369,27 @@ class MPxDetectorBase:
                 data_slice = slice(
                     chunk_firsts[0], int(chunk_firsts[-1] + chunk_counts[-1])
                 )
-                trainids = np.repeat(
-                    np.arange(first_tid, last_tid + 1, dtype=np.uint64),
-                    chunk_counts.astype(np.intp),
-                )
-                pulse_id = f.file['/INSTRUMENT/{}/{}/pulseId'.format(source, group)][
-                    data_slice
-                ]
-                # Raw files have a spurious extra dimension
-                if pulse_id.ndim >= 2 and pulse_id.shape[1] == 1:
-                    pulse_id = pulse_id[:, 0]
+
+                inner_ids = get_inner_ids(f, data_slice, inner_index)
 
                 if isinstance(pulses, by_id):
+                    if inner_index == 'pulseId':
+                        pulse_id = inner_ids
+                    else:
+                        pulse_id = get_inner_ids(f, data_slice, 'pulseId')
                     positions = self._select_pulse_ids(pulses, pulse_id)
                 else:  # by_index
                     positions = self._select_pulse_indices(
                         pulses, chunk_firsts - data_slice.start, chunk_counts
                     )
 
-                trainids = trainids[positions]
-                pulse_id = pulse_id[positions]
-                index = pd.MultiIndex.from_arrays(
-                    [trainids, pulse_id], names=['train', 'pulse']
+                trainids = np.repeat(
+                    np.arange(first_tid, last_tid + 1, dtype=np.uint64),
+                    chunk_counts.astype(np.intp),
                 )
+                index = self._make_image_index(
+                    trainids, inner_ids, inner_index[:-2]
+                )[positions]
 
                 if isinstance(positions, slice):
                     data_positions = slice(
@@ -297,9 +405,17 @@ class MPxDetectorBase:
                         # https://github.com/h5py/h5py/issues/1169
                         data_positions = slice(0, 0)
 
-                data = f.file[data_path][data_positions]
+                dset = f.file[data_path]
+                if dset.ndim >= 2 and dset.shape[1] == 1:
+                    # Ensure ROI applies to pixel dimensions, not the extra
+                    # dim in raw data (except AGIPD, where it is data/gain)
+                    sel_args = (data_positions, np.s_[:]) + roi
+                else:
+                    sel_args = (data_positions,) + roi
 
-                arr = _guess_axes(data, index, unstack_pulses)
+                data = f.file[data_path][sel_args]
+
+                arr = self._guess_axes(data, index, unstack_pulses)
 
                 seq_arrays.append(arr)
 
@@ -320,7 +436,9 @@ class MPxDetectorBase:
             dim=('train' if unstack_pulses else 'train_pulse'),
         )
 
-    def get_array(self, key, pulses=np.s_[:], unstack_pulses=True):
+    def get_array(self, key, pulses=np.s_[:], unstack_pulses=True, *,
+                  fill_value=None, subtrain_index='pulseId', roi=(),
+                  astype=None):
         """Get a labelled array of detector data
 
         Parameters
@@ -331,27 +449,52 @@ class MPxDetectorBase:
         pulses: slice, array, by_id or by_index
           Select the pulses to include from each train. by_id selects by pulse
           ID, by_index by index within the data being read. The default includes
-          all pulses. Only used for per-train data.
+          all pulses. Only used for per-pulse data.
         unstack_pulses: bool
           Whether to separate train and pulse dimensions.
+        fill_value: int or float, optional
+          Value to use for missing values. If None (default) the fill value is 0
+          for integers and np.nan for floats.
+        subtrain_index: str
+          Specify 'pulseId' (default) or 'cellId' to label the frames recorded
+          within each train. Pulse ID should allow this data to be matched with
+          other devices, but depends on how the detector was manually configured
+          when the data was taken. Cell ID refers to the memory cell used for
+          that frame in the detector hardware.
+        roi: tuple
+          Specify e.g. ``np.s_[10:60, 100:200]`` to select pixels within each
+          module when reading data. The selection is applied to each individual
+          module, so it may only be useful when working with a single module.
+          For AGIPD raw data, each module records a frame as a 3D array with 2
+          entries on the first dimension, for data & gain information, so
+          ``roi=np.s_[0]`` will select only the data part of each frame.
+        astype: Type
+          data type of the output array. If None (default) the dtype matches the
+          input array dtype
         """
-        pulses = _check_pulse_selection(pulses)
+        if subtrain_index not in {'pulseId', 'cellId'}:
+            raise ValueError("subtrain_index must be 'pulseId' or 'cellId'")
+        if not isinstance(roi, tuple):
+            roi = (roi,)
 
-        arrays = []
-        modnos = []
-        for modno, source in sorted(self.modno_to_source.items()):
-            # At present, all the per-pulse data is stored in the 'image' key.
-            # If that changes, this check will need to change as well.
-            if key.startswith('image.'):
+        if key.startswith('image.'):
+            pulses = _check_pulse_selection(pulses)
+
+            arrays, modnos = [], []
+            for modno, source in sorted(self.modno_to_source.items()):
                 arrays.append(self._get_module_pulse_data(
-                    source, key, pulses, unstack_pulses))
-            else:
-                arrays.append(self.data.get_array(source, key))
-            modnos.append(modno)
+                    source, key, pulses, unstack_pulses, subtrain_index, roi=roi
+                ))
+                modnos.append(modno)
 
-        return xarray.concat(arrays, pd.Index(modnos, name='module'))
+            return self._concat(arrays, modnos, fill_value, astype)
+        else:
+            return super().get_array(
+                key, fill_value=fill_value, roi=roi, astype=astype
+            )
 
-    def get_dask_array(self, key, subtrain_index='pulseId'):
+    def get_dask_array(self, key, subtrain_index='pulseId', fill_value=None,
+                       astype=None):
         """Get a labelled Dask array of detector data
 
         Dask does lazy, parallelised computing, and can work with large data
@@ -363,12 +506,18 @@ class MPxDetectorBase:
 
         key: str
           The data to get, e.g. 'image.data' for pixel values.
-        subtrain_index: str
+        subtrain_index: str, optional
           Specify 'pulseId' (default) or 'cellId' to label the frames recorded
           within each train. Pulse ID should allow this data to be matched with
           other devices, but depends on how the detector was manually configured
           when the data was taken. Cell ID refers to the memory cell used for
           that frame in the detector hardware.
+        fill_value: int or float, optional
+          Value to use for missing values. If None (default) the fill value is 0
+          for integers and np.nan for floats.
+        astype: Type, optional
+          data type of the output array. If None (default) the dtype matches the
+          input array dtype
         """
         if subtrain_index not in {'pulseId', 'cellId'}:
             raise ValueError("subtrain_index must be 'pulseId' or 'cellId'")
@@ -389,14 +538,17 @@ class MPxDetectorBase:
 
                 mod_arr = mod_arr.rename({'trainId': 'train_pulse'})
 
-                mod_arr.coords['train_pulse'] = pd.MultiIndex.from_arrays(
-                    [mod_arr.coords['train_pulse'].values, inner_ix.values],
-                    names=['trainId', subtrain_index]
-                )
+                mod_arr.coords['train_pulse'] = self._make_image_index(
+                    mod_arr.coords['train_pulse'].values, inner_ix.values,
+                    inner_name=subtrain_index,
+                ).set_names('trainId', level=0)
+                # This uses 'trainId' where a concrete array from the same class
+                # uses 'train'. I didn't notice that inconsistency when I
+                # introduced it, and now code may be relying on each name.
 
             arrays.append(mod_arr)
 
-        return xarray.concat(arrays, pd.Index(modnos, name='module'))
+        return self._concat(arrays, modnos, fill_value, astype)
 
     def trains(self, pulses=np.s_[:], require_all=True):
         """Iterate over trains for detector data.
@@ -604,9 +756,7 @@ class MPxDetectorTrainIterator:
             positions = self._select_pulse_indices(count)
         pulse_ids = pulse_ids[positions]
         train_ids = np.array([tid] * len(pulse_ids), dtype=np.uint64)
-        train_pulse_ids = pd.MultiIndex.from_arrays(
-            [train_ids, pulse_ids], names=['train', 'pulse']
-        )
+        train_pulse_ids = self.data._make_image_index(train_ids, pulse_ids)
 
         if isinstance(positions, slice):
             data_positions = slice(
@@ -618,7 +768,7 @@ class MPxDetectorTrainIterator:
             # h5py fancy indexing needs a list, not an ndarray
             data_positions = list(first + positions)
 
-        return _guess_axes(ds[data_positions], train_pulse_ids, unstack_pulses=True)
+        return self.data._guess_axes(ds[data_positions], train_pulse_ids, unstack_pulses=True)
 
     def _select_pulse_ids(self, pulse_ids):
         """Select pulses by ID
@@ -692,14 +842,14 @@ class MPxDetectorTrainIterator:
             yield tid, self._assemble_data(tid)
 
 
-class AGIPD1M(MPxDetectorBase):
+class AGIPD1M(XtdfDetectorBase):
     """An interface to AGIPD-1M data.
 
     Parameters
     ----------
 
     data: DataCollection
-      A data collection, e.g. from RunDirectory.
+      A data collection, e.g. from :func:`.RunDirectory`.
     modules: set of ints, optional
       Detector module numbers to use. By default, all available modules
       are used.
@@ -709,18 +859,29 @@ class AGIPD1M(MPxDetectorBase):
     min_modules: int
       Include trains where at least n modules have data. Default is 1.
     """
-    _source_re = re.compile(r'(?P<detname>(.+)_AGIPD1M(.*))/DET/(\d+)CH')
+    _source_re = re.compile(r'(?P<detname>.+_AGIPD1M.*)/DET/(?P<modno>\d+)CH')
     module_shape = (512, 128)
 
 
-class DSSC1M(MPxDetectorBase):
+class AGIPD500K(XtdfDetectorBase):
+    """An interface to AGIPD-500K data
+
+    Detector names are like 'HED_DET_AGIPD500K2G', otherwise this is identical
+    to :class:`AGIPD1M`.
+    """
+    _source_re = re.compile(r'(?P<detname>.+_AGIPD500K.*)/DET/(?P<modno>\d+)CH')
+    module_shape = (512, 128)
+    n_modules = 8
+
+
+class DSSC1M(XtdfDetectorBase):
     """An interface to DSSC-1M data.
 
     Parameters
     ----------
 
     data: DataCollection
-      A data collection, e.g. from RunDirectory.
+      A data collection, e.g. from :func:`.RunDirectory`.
     modules: set of ints, optional
       Detector module numbers to use. By default, all available modules
       are used.
@@ -730,18 +891,18 @@ class DSSC1M(MPxDetectorBase):
     min_modules: int
       Include trains where at least n modules have data. Default is 1.
     """
-    _source_re = re.compile(r'(?P<detname>(.+)_DSSC1M(.*))/DET/(\d+)CH')
+    _source_re = re.compile(r'(?P<detname>.+_DSSC1M.*)/DET/(?P<modno>\d+)CH')
     module_shape = (128, 512)
 
 
-class LPD1M(MPxDetectorBase):
+class LPD1M(XtdfDetectorBase):
     """An interface to LPD-1M data.
 
     Parameters
     ----------
 
     data: DataCollection
-      A data collection, e.g. from RunDirectory.
+      A data collection, e.g. from :func:`.RunDirectory`.
     modules: set of ints, optional
       Detector module numbers to use. By default, all available modules
       are used.
@@ -750,9 +911,161 @@ class LPD1M(MPxDetectorBase):
       if the dataset includes more than one LPD detector.
     min_modules: int
       Include trains where at least n modules have data. Default is 1.
+    parallel_gain: bool
+      Set to True to read this data as parallel gain data, where high, medium
+      and low gain data are stored sequentially within each train. This will
+      repeat the pulse & cell IDs from the first 1/3 of each train, and add gain
+      stage labels from 0 (high-gain) to 2 (low-gain).
     """
-    _source_re = re.compile(r'(?P<detname>(.+)_LPD1M(.*))/DET/(\d+)CH')
+    _source_re = re.compile(r'(?P<detname>.+_LPD1M.*)/DET/(?P<modno>\d+)CH')
     module_shape = (256, 256)
+
+    def __init__(self, data: DataCollection, detector_name=None, modules=None,
+                 *, min_modules=1, parallel_gain=False):
+        super().__init__(data, detector_name, modules, min_modules=min_modules)
+
+        self.parallel_gain = parallel_gain
+        if parallel_gain:
+            if ((self.frame_counts % 3) != 0).any():
+                raise ValueError(
+                    "parallel_gain=True needs the frames in each train to be divisible by 3"
+                )
+
+    def _make_image_index(self, tids, inner_ids, inner_name='pulse'):
+        if not self.parallel_gain:
+            return super()._make_image_index(tids, inner_ids, inner_name)
+
+        # In 'parallel gain' mode, the first 1/3 of pulse/cell IDs in each train
+        # are valid, but the remaining 2/3 are junk. So we'll repeat the valid
+        # ones 3 times (in inner_ids_fixed). At the same time, we make a gain
+        # stage index (0-2), so each frame has a unique entry in the MultiIndex
+        # (train ID, gain, pulse/cell ID)
+        gain = np.zeros_like(inner_ids, dtype=np.uint8)
+        inner_ids_fixed = np.zeros_like(inner_ids)
+
+        _, firsts, counts = np.unique(tids, return_index=True, return_counts=True)
+        for ix, frames in zip(firsts, counts):  # Iterate through trains
+            n_per_gain_stage = int(frames // 3)
+            train_inner_ids = inner_ids[ix: ix + n_per_gain_stage]
+            for stage in range(3):
+                start = ix + (stage * n_per_gain_stage)
+                end = start + n_per_gain_stage
+                gain[start:end] = stage
+                inner_ids_fixed[start:end] = train_inner_ids
+
+        return pd.MultiIndex.from_arrays(
+            [tids, gain, inner_ids_fixed], names=['train', 'gain', inner_name]
+        )
+
+
+class JUNGFRAU(MultimodDetectorBase):
+    """An interface to JUNGFRAU data.
+
+    Parameters
+    ----------
+
+    data: DataCollection
+      A data collection, e.g. from :func:`.RunDirectory`.
+    modules: set of ints, optional
+      Detector module numbers to use. By default, all available modules
+      are used.
+    detector_name: str, optional
+      Name of a detector, e.g. 'SPB_IRDA_JNGFR'. This is only needed
+      if the dataset includes more than one JUNGFRAU detector.
+    min_modules: int
+      Include trains where at least n modules have data. Default is 1.
+    """
+    # We appear to have a few different formats for source names:
+    # SPB_IRDA_JNGFR/DET/MODULE_1:daqOutput  (e.g. in p 2566, r 61)
+    # SPB_IRDA_JF4M/DET/JNGFR03:daqOutput    (e.g. in p 2732, r 12)
+    # FXE_XAD_JF1M/DET/RECEIVER-1
+    _source_re = re.compile(
+        r'(?P<detname>.+_(JNGFR|JF[14]M))/DET/(MODULE_|RECEIVER-|JNGFR)(?P<modno>\d+)'
+    )
+    _main_data_key = 'data.adc'
+    module_shape = (512, 1024)
+
+    @staticmethod
+    def _label_dims(arr):
+        # Label dimensions to match the AGIPD/DSSC/LPD data access
+        ndim_pertrain = arr.ndim
+        if 'trainId' in arr.dims:
+            arr = arr.rename({'trainId': 'train'})
+            ndim_pertrain = arr.ndim - 1
+
+        if ndim_pertrain == 4:
+            arr = arr.rename({
+                'dim_0': 'pulse', 'dim_1': 'slow_scan', 'dim_2': 'fast_scan'
+            })
+        elif ndim_pertrain == 2:
+            arr = arr.rename({'dim_0': 'pulse'})
+        return arr
+
+    def get_array(self, key, *, fill_value=None, roi=(), astype=None):
+        """Get a labelled array of detector data
+
+        Parameters
+        ----------
+
+        key: str
+          The data to get, e.g. 'data.adc' for pixel values.
+        fill_value: int or float, optional
+            Value to use for missing values. If None (default) the fill value
+            is 0 for integers and np.nan for floats.
+        roi: tuple
+          Specify e.g. ``np.s_[:, 10:60, 100:200]`` to select data within each
+          module & each train when reading data. The first dimension is pulses,
+          then there are two pixel dimensions. The same selection is applied
+          to data from each module, so selecting pixels may only make sense if
+          you're using a single module.
+        astype: Type
+          data type of the output array. If None (default) the dtype matches the
+          input array dtype
+        """
+        arr = super().get_array(key, fill_value=fill_value, roi=roi, astype=astype)
+        return self._label_dims(arr)
+
+    def get_dask_array(self, key, fill_value=None, astype=None):
+        """Get a labelled Dask array of detector data
+
+        Dask does lazy, parallelised computing, and can work with large data
+        volumes. This method doesn't immediately load the data: that only
+        happens once you trigger a computation.
+
+        Parameters
+        ----------
+
+        key: str
+          The data to get, e.g. 'data.adc' for pixel values.
+        fill_value: int or float, optional
+          Value to use for missing values. If None (default) the fill value
+          is 0 for integers and np.nan for floats.
+        astype: Type
+          data type of the output array. If None (default) the dtype matches the
+          input array dtype
+        """
+        arr = super().get_dask_array(key, fill_value=fill_value, astype=astype)
+        return self._label_dims(arr)
+
+    def trains(self, require_all=True):
+        """Iterate over trains for detector data.
+
+        Parameters
+        ----------
+
+        require_all: bool
+          If True (default), skip trains where any of the selected detector
+          modules are missing data.
+
+        Yields
+        ------
+
+        train_data: dict
+          A dictionary mapping key names (e.g. 'data.adc') to labelled
+          arrays.
+        """
+        for tid, d in super().trains(require_all=require_all):
+            yield tid, {k: self._label_dims(a) for (k, a) in d.items()}
 
 
 def identify_multimod_detectors(