+++ extra_data/tests/test_components.py
@@ -6,7 +6,9 @@ import pytest
 from testpath import assert_isfile
 
 from extra_data.reader import RunDirectory, H5File, by_id, by_index
-from extra_data.components import AGIPD1M, LPD1M, identify_multimod_detectors
+from extra_data.components import (
+    AGIPD1M, DSSC1M, LPD1M, JUNGFRAU, identify_multimod_detectors,
+)
 
 
 def test_get_array(mock_fxe_raw_run):
@@ -15,12 +17,22 @@ def test_get_array(mock_fxe_raw_run):
     assert det.detector_name == 'FXE_DET_LPD1M-1'
 
     arr = det.get_array('image.data')
+    assert arr.dtype == np.uint16
     assert arr.shape == (16, 3, 128, 256, 256)
     assert arr.dims == ('module', 'train', 'pulse', 'slow_scan', 'fast_scan')
 
     arr = det.get_array('image.data', pulses=by_index[:10], unstack_pulses=False)
     assert arr.shape == (16, 30, 256, 256)
+    assert arr.dtype == np.uint16
     assert arr.dims == ('module', 'train_pulse', 'slow_scan', 'fast_scan')
+
+    # fill value
+    with pytest.raises(ValueError):
+        det.get_array('image.data', fill_value=np.nan)
+    
+    arr = det.get_array('image.data', astype=np.float32)
+    assert arr.dtype == np.float32
+
 
 def test_get_array_pulse_id(mock_fxe_raw_run):
     run = RunDirectory(mock_fxe_raw_run)
@@ -42,6 +54,18 @@ def test_get_array_pulse_id(mock_fxe_raw_run):
     arr = det.get_array('image.data', pulses=by_id[[1, 7, 22, 23]])
     assert arr.shape == (16, 3, 4, 256, 256)
     assert list(arr.coords['pulse']) == [1, 7, 22, 23]
+
+
+def test_get_array_with_cell_ids(mock_fxe_raw_run):
+    run = RunDirectory(mock_fxe_raw_run)
+    det = LPD1M(run.select_trains(by_index[:3]))
+    arr = det.get_array('image.data', subtrain_index='cellId')
+    assert arr.shape == (16, 3, 128, 256, 256)
+    assert arr.dims == ('module', 'train', 'cell', 'slow_scan', 'fast_scan')
+
+    arr = det.get_array('image.data', pulses=by_id[0], subtrain_index='cellId')
+    assert arr.shape == (16, 3, 1, 256, 256)
+    assert (arr.coords['cell'] == 0).all()
 
 
 def test_get_array_pulse_indexes(mock_fxe_raw_run):
@@ -109,13 +133,59 @@ def test_get_array_pulse_indexes_reduced_data(mock_reduced_spb_proc_run):
     arr = det.get_array('image.data', pulses=[1, 7, 15, 23])
     assert np.isin(arr.coords['pulse'], [1, 7, 15, 23]).all()
 
+
+def test_get_array_roi(mock_fxe_raw_run):
+    run = RunDirectory(mock_fxe_raw_run)
+    det = LPD1M(run.select_trains(by_index[:3]))
+    assert det.detector_name == 'FXE_DET_LPD1M-1'
+
+    arr = det.get_array('image.data', roi=np.s_[10:60, 100:200])
+    assert arr.shape == (16, 3, 128, 50, 100)
+    assert arr.dims == ('module', 'train', 'pulse', 'slow_scan', 'fast_scan')
+
+
+def test_get_array_roi_dssc(mock_scs_run):
+    run = RunDirectory(mock_scs_run)
+    det = DSSC1M(run, modules=[3])
+
+    arr = det.get_array('image.data', roi=np.s_[20:25, 40:52])
+    assert arr.shape == (1, 128, 64, 5, 12)
+
+
+def test_get_array_lpd_parallelgain(mock_lpd_parallelgain_run):
+    run = RunDirectory(mock_lpd_parallelgain_run)
+    det = LPD1M(run.select_trains(by_index[:2]), parallel_gain=True)
+    assert det.detector_name == 'FXE_DET_LPD1M-1'
+
+    arr = det.get_array('image.data')
+    assert arr.shape == (16, 2, 3, 100, 256, 256)
+    assert arr.dims == ('module', 'train', 'gain', 'pulse', 'slow_scan', 'fast_scan')
+    np.testing.assert_array_equal(arr.coords['gain'], np.arange(3))
+    np.testing.assert_array_equal(arr.coords['pulse'], np.arange(100))
+
+
+def test_get_array_jungfrau(mock_jungfrau_run):
+    run = RunDirectory(mock_jungfrau_run)
+    jf = JUNGFRAU(run.select_trains(by_index[:2]))
+    assert jf.detector_name == 'SPB_IRDA_JF4M'
+
+    arr = jf.get_array('data.adc')
+    assert arr.shape == (8, 2, 16, 512, 1024)
+    assert arr.dims == ('module', 'train', 'pulse', 'slow_scan', 'fast_scan')
+    np.testing.assert_array_equal(arr.coords['train'], [10000, 10001])
+
+    arr = jf.get_array('data.adc', astype=np.float32)
+    assert arr.dtype == np.float32
+
+
 def test_get_dask_array(mock_fxe_raw_run):
     run = RunDirectory(mock_fxe_raw_run)
     det = LPD1M(run)
-    arr = det.get_dask_array('image.data')
+    arr = det.get_dask_array('image.data', fill_value=42)
 
     assert isinstance(arr.data, da.Array)
     assert arr.shape == (16, 480 * 128, 1, 256, 256)
+    assert arr.dtype == np.uint16
     assert arr.dims == ('module', 'train_pulse', 'dim_0', 'dim_1', 'dim_2')
     np.testing.assert_array_equal(arr.coords['module'], np.arange(16))
     np.testing.assert_array_equal(
@@ -139,6 +209,28 @@ def test_get_dask_array_reduced_data(mock_reduced_spb_proc_run):
     np.testing.assert_array_equal(arr.coords['module'], np.arange(16))
     assert np.isin(arr.coords['trainId'], np.arange(10000, 10480)).all()
     assert np.isin(arr.coords['pulseId'], np.arange(0, 20)).all()
+
+
+def test_get_dask_array_lpd_parallelgain(mock_lpd_parallelgain_run):
+    run = RunDirectory(mock_lpd_parallelgain_run)
+    det = LPD1M(run.select_trains(by_index[:2]), parallel_gain=True)
+    assert det.detector_name == 'FXE_DET_LPD1M-1'
+
+    arr = det.get_dask_array('image.data')
+    assert arr.shape == (16, 2 * 3 * 100, 1, 256, 256)
+    assert arr.dims[:2] == ('module', 'train_pulse')
+    np.testing.assert_array_equal(arr.coords['pulseId'], np.tile(np.arange(100), 6))
+
+
+def test_get_dask_array_jungfrau(mock_jungfrau_run):
+    run = RunDirectory(mock_jungfrau_run)
+    jf = JUNGFRAU(run)
+    assert jf.detector_name == 'SPB_IRDA_JF4M'
+
+    arr = jf.get_dask_array('data.adc')
+    assert arr.shape == (8, 100, 16, 512, 1024)
+    assert arr.dims == ('module', 'train', 'pulse', 'slow_scan', 'fast_scan')
+    np.testing.assert_array_equal(arr.coords['train'], np.arange(10000, 10100))
 
 
 def test_iterate(mock_fxe_raw_run):
@@ -188,6 +280,27 @@ def test_iterate_pulse_index(mock_fxe_raw_run):
     tid, d = next(iter(det.trains(pulses=by_index[[1, 7, 22, 23]])))
     assert d['image.data'].shape == (16, 1, 4, 256, 256)
     assert list(d['image.data'].coords['pulse']) == [1, 7, 22, 23]
+
+
+def test_iterate_lpd_parallel_gain(mock_lpd_parallelgain_run):
+    run = RunDirectory(mock_lpd_parallelgain_run)
+    det = LPD1M(run.select_trains(by_index[:3]), parallel_gain=True)
+
+    tid, d = next(iter(det.trains()))
+    assert d['image.data'].shape == (16, 1, 3, 100, 256, 256)
+    assert d['image.data'].dims == \
+           ('module', 'train', 'gain', 'pulse', 'slow_scan', 'fast_scan')
+
+
+def test_iterate_jungfrau(mock_jungfrau_run):
+    run = RunDirectory(mock_jungfrau_run)
+    jf = JUNGFRAU(run)
+
+    tid, d = next(iter(jf.trains()))
+    assert tid == 10000
+    assert d['data.adc'].shape == (8, 16, 512, 1024)
+    assert d['data.adc'].dims == ('module', 'pulse', 'slow_scan', 'fast_scan')
+
 
 def test_write_virtual_cxi(mock_spb_proc_run, tmpdir):
     run = RunDirectory(mock_spb_proc_run)